Introduce AccessRequest CRD
Sveltos might need to deploy different agents in managed cluster. For
instance classifier-agent is deployed in each managed cluster and
classifier-agent might be started with option to update ClassifierReport
in management cluster.

When any of those agents need a kubeconfig, an AccessRequest can be
created and when Kubeconfig is ready, Status will contain it.