Introduce RoleRequest CRD
In a multi-tenancy environment, the platform admin can grant subset
of permissions to tenant admin by creating a RoleRequest instance.
- ClusterSelector is a pure Kubernetes label selector and identifies in
which clusters those permissions will be granted;
- RoleRefs references ConfigMaps each containing one or more
(Cluster)Roles;
- admin is the name of the tenant admin to whome permissions are
granted.

After RoleRequest is created, a tenant admin will be able to create
ClusterProfile and the resources (helm charts or YAML resources)
contained in the ClusterProfile will be deployed if tenant admin has
right set of permissions.