Move ValidateObjectForUpdate to libsveltos
ClusterProfile and RoleRequest both references ConfigMaps/Secrets
containing policies that need to be deployed in matching clusters.

All resources contained in a referenced ConfigMap/Secret are deployed in
a matching managed cluster.

Because same resource contained in a ConfigMap/Secret, can also be
contained in other referenced ConfigMaps/Secrets, before deploying a
resource in a matching cluster following validation is needed:
1) validate resource is not already present;
2) if resource is already present, validate that it was contained by the
same referenced ConfigMap/Secret trying to deploy it now.

This code was originally in sveltos-manager. Given RoleRequest needs
same validation, moving it (along with Labels/Annotation) in libsveltos.