Cluster annotation for addon constraint
Cluster addons are deployed by addon controller.
Cluster addon constraints are loaded by addon-constraint controller.
When new cluster is created it might match both a ClusterProfile and an
AddonConstrain.
Matching a ClusterProfile means addons need to be deployed (by addon
controller).
Matching an AddonConstraint means some validations are defined and any
addon deployed in this cluster should satisfy those validations.

We need a sistem that allows addon constraints to be loaded before any
addon is deployed.
Addon-constraint controller will add this annotation to any cluster it
is aware of (if addon-constrain controller is aware of a cluster we can
assume validations for such cluster are loaded).

Addon controller won't deploy any addons till this annotation is set on
a cluster.