Introducing sharding for sveltos controller
This PR introduces:
1. the annotation to add to clusters to shard them
2. utility functions needed by controllers to get clusters filtering
by shard and detecting whether a cluster belongs to a given shard

In general, a cluster with the annotation

```
sharding.projectsveltos.io/key: <shard-value>
```

will only be managed by the controller started explicitly for that
shard.

A cluster with no annotation will be managed by the default controller.