Extend EventSource
Resources of different GroupVersionKinds can be selected using
ResourceSelectors field.
ResourceSelectors is an array of ResourceSelector. Each ResourceSelector
select resources of a given GroupVersionKind and filter them by:
1. namespace
2. label
3. any other field or combination of fields by Lua function.

AggregatedSelection is an optional field that allows to further
filter resources selected with ResourceSelectors.
AggregatedSelection contains a Lua function that will be used to
further select a subset of the resources that have already been
selected using the ResourceSelector field.
The function will receive the array of resources selected by ResourceSelectors.
If this field is not specified, all resources selected by the ResourceSelector
field will be considered. This field allows to perform more complex filtering
or selection operations on the resources, looking at all resources together.
This can be useful for more sophisticated tasks, such as identifying resources
that are related to each other or that have similar properties.