Classifier CRD
This PR aligns Classifier instance with EventSource instance.
When classifying a Cluster based on deployed resources, Classifier
the field DeployedResourceConstraint has:

- slice of ResourceSelectors each selecting resources of a given
group version kind, filtering based on namespace, labels or a Lua
script;
- AggregatedClassification which is an optional Lua script that looks
at resources selected by ResourceSelectors all together and returns
a boolean indicating whether cluster is a match.

If AggregatedClassification field is specified, a Cluster is a match
for Classifier if AggregatedClassification returns true.
If AggregatedClassification is not specified, a Cluster is match for
Classifier if each ResourceSelector returns at least one match.