Deployed resource validation
A Kubernetes resource can be deployed in a managed cluster only by
a ClusterProfile/Profile instance

Before this PR, if two ClusterProfile/Profile instances:
1. matched the same cluster
2. referenced the same ConfigMap

the resources contained in the ConfigMap were successfully deployed
and neither of the ClusterProfile/Profile instance reported an error.

After this PR, the resources contained in the ConfigMap will still be
successfully deployed. But while the first ClusterProfile/Profile will
succeed, the second one will report a conflict with the first one.

This behavior change is implemented as part of introducing the concept
of tier in ClusterProfile/Profile.
Allowing only one ClusterProfile/Profile instance to deploy a given
Kubernetes resources, allows to handle the scenario where different
ClusterProfile/Profile instances with different tiers want to deploy
the same resource (and let the profile with lower tier win the conflict).