(feat) add patches to ResourceSummary
When creating Profile/ClusterProfile with syncMode set to
`ContinuousWithDriftDetection` user will have the ability
to specify set of paths for certain resources that need to
be ignored during configuration drift evaluation.
User can use the field `DriftExclusions`.

This is helpful for instance when a Deployment is deployed
by Sveltos along an autoscaler. Autoscaler will scale `replicas`
up and down. So we don't want a replicas change to be seen as
a configuration drift.

addon-manager will pass resources that need to be watched for
configuration drift and will transform DriftExclusions into a
slices of `Patch`.
Operation is always set to remove as the goal of a DriftExclusion
is to not consider, so to remove, a path during configuration drift
evaluation.

Patch (as the whole ResourceSummary instance) will be consumed
by drift-detection-manager.