(feat) Add name to ResourceSelector
In several situation, an event is identified by the creation of a specific
resource.
Add `name` to ResourceSelector to better handle this scenarion.

Before this PR, to watch for a specific resource, we had to write Lua code
in the `evaluate` field.
With this PR, that is not needed anymore.