(feat) Selectively onboard CAPI clusters
Deploying Sveltos on a management cluster with hundreds of existing CAPI clusters
results in all clusters being onboarded immediately. This can lead to a rapid creation
of hundreds of sveltos-agent deployments in the management cluster, potentially causing
performance issues.  Therefore, a gradual onboarding mechanism is preferable.

This pull request introduces such a mechanism.  By using an annotation, Sveltos can selectively
identify which CAPI clusters to manage. When this annotation is specified, Sveltos will only onboard
those CAPI clusters that have it.

This feature allows for gradual onboarding, which is particularly beneficial in management clusters
with a large number of pre-existing CAPI clusters.