(feat) add ConfigurationGroup and ConfigurationBundle CRDs
A ConfigurationBundle contains a list of resources that need
to be deployed on a given managed cluster.

A ConfigurationGroup contains a list of ConfigurationBundles with
the associated hash.

ConfigurationGroup does not have a finalizer. Only Sveltos component
in the management cluster create and delete ConfigurationGroups.
ConfigurationGroup has a field indicating whether the content referenced
by ConfigurationGroup should be deployed or all deployed because of a
ConfigurationGroup should be withdrawn.

Agent will watch for ConfigurationGroup and then download the referenced
ConfigurationBundles.

The pullmode package includes all the common logic multiple Sveltos agents
in the management cluster and the agent in the managed cluster need.