(feat) add helper methods for ConfigMap with EventSource/HealthCheck
When sveltos-agent is deployed in the management cluster, EventSource and HealthCheck
instances are not copied to the managed cluster.
A per cluster ConfigMap is created containing the EventSource and HealthCheck instances
meant to be evaluated in a given cluster.
This PR add helper methods to identify if an entry was created because of an EventTrigger
instance or a ClusterHealthCheck instance.