(feat) add annotations for specifying profile owner
When Sveltos deploys YAML/JSON (via PolicyRefs section) used to add
as OwnerReference the profile (ClusterProfile/Profile) that caused that
resource to be deployed.

If Sveltos is deploying such resources to the management cluster or any
other cluster where Sveltos is also present, that will fail as
ClusterProfile is a cluster-wide Custom Resource Definition (CRD) and cannot
be set as OwnerReference for namespaced resources.

Therefore, Sveltos needs to avoid using OwnerReference in such scenarios.
This PR sets annotations on deployed resources instead of ownerReference.