(bug) use retry on conflict when removing staged label
ConfigurationBundles are created and then Status is updated with hash.
There is a chance that when CommitStagedResourcesForDeployment is called
in the cache we have old instances of ConfigurationBundle (before the status update).

The CommitStagedResourcesForDeployment tries to remove the staged label
from each ConfigurationBundle instance and that might fail because cache
is not updated yet.

This PR wraps the staged label remove around a retry on conflict.