(chore) Add new label for EventReport
This change introduces a new label to EventReport resources.
For managed clusters in pull-mode, sveltos-applier collects
EventReports from managed cluster and copies them to the
management cluster.

To enable the event-manager to efficiently monitor and react
to these specific EventReport instances, sveltos-applier
automatically attach a this label during the copying process.