Add ServiceAccount information to ConfigurationGroup
When set, Sveltos-applier will impersonate before applying
resources to the managed cluster

This is to support RoleRequests in pull mode.
When RoleRequest is created, access-manager will deploy, via sveltos-applier,
a ServiceAccount and associated ClusterRole/Role.

When the labels `projectsveltos.io/serviceaccount-name`and
`projectsveltos.io/serviceaccount-namespace` are set on a ClusterProfile/Profile
instance, this information is passed to sveltos-applier via ConfigurationGroup.

Sveltos-applier will then impersonate that ServiceAccount before deploying