(bug) Cluster predicates did not react to Cluster conditions changing
__ClusterControlPlaneInitializedCondition__ is true when the CAPI Cluster's
control plane is functional enough to accept requests. This information is
usually used as a signal for starting all the provisioning operations
that depends on a functional API server, but do not require a full HA control
plane to exists.

Sveltos must react when this is set to true. And Cluster predicate was not
reacting to it.