EventReport: mark matchingResources with listType=atomic
A EventSource starts a watcher on specified resources. Any resource
change triggers the EventSource to be reprocessed, which discovers
all matching resources and updates the EventReport.

When collectResources is set to false, an EventSource might be processed
multiple times even if the list of matching resources hasn't changed.
Any EventReport update causes event-manager to reprocess it on the control
cluster.

To prevent redundant updates to the EventReport the update
can be skipped entirely, provided the matching resources are sorted
and maintained in a sorted state within the report.